add comments on controller files

+++ src/main/java/com/example/demo/back/controller/MessageContoller.java
@@ -1,7 +1,6 @@
 package com.example.demo.back.controller;
-import com.example.demo.back.model.Channel;
+
 import com.example.demo.back.model.Message;
-import com.example.demo.back.model.User;
 import com.example.demo.back.service.MessageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,16 +11,29 @@ import java.util.*;
 
 @RestController
 public class MessageContoller {
+
     @Autowired
     MessageService messageService;
+
+    /**
+     * Récupère tous les messages.
+     *
+     * @return Une liste de tous les messages.
+     */
     @GetMapping("messages")
-    public List<Message> findAll(){
+    public List<Message> findAll() {
         return messageService.findAll();
     }
 
+    /**
+     * Ajoute un nouveau message.
+     *
+     * @param message Le message à ajouter.
+     * @return réponse HTTP indiquant le succès ou l'échec de l'ajout.
+     */
     @PostMapping("messages")
-    public ResponseEntity<Message> login(@RequestBody Message message){
-        if(message.getContent().isEmpty())
+    public ResponseEntity<?> addMessage(@RequestBody Message message) {
+        if (message.getContent().isEmpty())
             return ResponseEntity
                     .badRequest()
                     .build();
@@ -30,40 +42,60 @@ public class MessageContoller {
             return ResponseEntity.status(HttpStatus.CREATED).build();
         }
     }
+
+    /**
+     * Récupère un message par son identifiant.
+     *
+     * @param id L'identifiant du message à récupérer.
+     * @return une Réponse HTTP contenant le message s'il est trouvé, sinon retourne une réponse NotFound.
+     */
     @GetMapping("messages/{id}")
-    public ResponseEntity<Message> findById(@PathVariable("id") Integer id){
+    public ResponseEntity<?> findById(@PathVariable("id") Integer id) {
         Optional<Message> opt = messageService.findById(id);
-        if(opt.isEmpty())
-            return ResponseEntity.notFound().build();
+        if (opt.isEmpty())
+            return ResponseEntity
+                    .notFound()
+                    .build();
         else
             return ResponseEntity.ok(opt.get());
     }
-    @DeleteMapping("messages/{id}")
-    public ResponseEntity<Message> delete(@PathVariable("id") Integer id,
-                                          @RequestBody Message message){
-        Optional<Message> opt = messageService.findById(id);
-        if (opt.isEmpty()){
-            return ResponseEntity
-                    .notFound()
-                    .build();
-        }    if (message.getId()!=id){
 
+    /**
+     * Met à jour un message existant.
+     *
+     * @param id      L'identifiant du message à mettre à jour.
+     * @param message Le nouveau contenu du message mis à jour.
+     * @return une Réponse HTTP indiquant le succès ou l'échec de la mise à jour.
+     */
+    @PutMapping("messages/{id}")
+    public ResponseEntity<?> update(@PathVariable("id") Integer id,
+                                    @RequestBody Message message) {
+        if (!message.getId().equals(id)) {
             return ResponseEntity
                     .notFound()
                     .build();
-        }else
-            messageService.delete(message.getId());
+        } else
+            messageService.update(message);
         return ResponseEntity.ok().build();
     }
-    @PutMapping("messages/{id}")
-    public ResponseEntity<Message> update(@PathVariable("id") Integer id,
-                                          @RequestBody Message message){
-       if (message.getId()!=id){
-           return ResponseEntity
-                   .notFound()
-                   .build();
-       }else
-        messageService.update(message);
+
+    /**
+     * Supprime un message existant.
+     *
+     * @param id      L'identifiant du message à supprimer.
+     * @param message Le message à supprimer.
+     * @return Réponse HTTP indiquant le succès ou l'échec de l'opération.
+     */
+    @DeleteMapping("messages/{id}")
+    public ResponseEntity<?> delete(@PathVariable("id") Integer id,
+                                    @RequestBody Message message) {
+        Optional<Message> opt = messageService.findById(id);
+        if (opt.isEmpty() && message.getId().equals(id)) { // message.getId() != id
+            return ResponseEntity
+                    .notFound()
+                    .build();
+        } else
+            messageService.delete(message.getId());
         return ResponseEntity.ok().build();
     }
